requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
celery==5.3.6
aio-pika==9.3.1
jinja2==3.1.2
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
